refactor: split campaigns.ts and resources.ts under the 800-line cap
Bring the two outsized dashboard route files back under the AGENTS.md
800-line guideline by extracting cohesive route groups into sibling
modules. Each extracted file exports a `registerXxxRoutes(router, ...)`
function the main file calls after constructing the router, so URL
paths, middleware composition, and request semantics stay identical.

campaigns.ts (1194 → 588 lines)
- New `campaigns-shared.ts`: campaignIdParamSchema, campaignRowSchema,
  attackRowSchema, lifecycleResponseSchema, respondToTransition,
  checkResourcesOrErrorResponse. Shared across the three campaign
  route files; lives in its own module so the lifecycle and attack
  files don't import back through `campaigns.ts` (cycle-free).
- New `campaigns-lifecycle.ts` (185 lines): /{id}/lifecycle action-enum
  route + five spec-named aliases (start/pause/resume/stop/cancel).
- New `campaigns-attacks.ts` (393 lines): POST/GET attacks,
  PATCH/DELETE /{id}/attacks/{attackId}. Cross-project guards and
  DAG validation preserved.

resources.ts (1322 → 679 lines)
- New `resources-shared.ts` (109 lines): enforceMultipartSizeLimit,
  passthroughObject, idParamSchema, uploadIdParamSchema,
  uploadPartParamSchema, tags, security, uploadStatusQuerySchema,
  RESOURCE_TYPES, MULTIPART_BODY_LIMIT_BYTES.
- New `resources-generic.ts` (295 lines): the wordlists/rulelists/
  masklists factory — list/create/get/upload/delete/download — now
  exported as `registerGenericResourceRoutes(router, prefix, table)`.
- New `resources-chunked-upload.ts` (347 lines): S3 multipart session
  routes — initiate/part/complete/abort/status. Raw octet-stream part
  route preserved (route declares the content type but handler still
  uses c.req.arrayBuffer() directly).

Every file is now under the 800-line cap (max: resources.ts at 679).
just ci-check green: 4/4 backend test phases pass including the
101-case DASHBOARD_CAMPAIGNS_TEST_ISOLATED run that exercises every
campaigns route after the split, and all dashboard-resources tests.
