+++ packages/backend/src/routes/dashboard/campaigns.ts
@@ -15,24 +15,20 @@ import {
   sharedResponse,
 } from '../../openapi/components.js'
 import {
-  createAttack,
   createCampaign,
   createCampaignWithAttacks,
-  deleteAttack,
   deleteCampaign,
-  getAttackById,
   getCampaignById,
   getCampaignTaskStats,
   listActiveAgentsByCampaign,
   listAttacks,
   listCampaigns,
-  transitionCampaign,
-  updateAttack,
   updateCampaign,
   validateCampaignDAG,
-  validateCampaignResources,
-  validateProposedDAG,
 } from '../../services/campaigns.js'
+import { registerCampaignAttackRoutes } from './campaigns-attacks.js'
+import { registerCampaignLifecycleRoutes } from './campaigns-lifecycle.js'
+import { attackRowSchema, campaignIdParamSchema, campaignRowSchema } from './campaigns-shared.js'
 
 // Use the shared dashboardOpenApiHonoOptions so every dashboard router
 // emits the same `{ error: { code: 'VALIDATION_ERROR', message } }`
@@ -74,21 +70,9 @@ const listCampaignsQuerySchema = z.object({
   offset: z.coerce.number().int().min(0).catch(0).default(0),
 })
 
-// Path-param schema: `id` is the campaign id everywhere on this
-// router. `z.coerce.number().int().positive()` replaces the inline
-// `Number.isInteger(id) || id <= 0` check that was repeated in every
-// handler; an invalid id surfaces as the defaultHook's 400
-// VALIDATION_ERROR envelope before the handler runs.
-const campaignIdParamSchema = z.object({
-  id: z.coerce.number().int().positive(),
-})
-
-// Passthrough schemas for response shapes that don't (yet) have a
-// canonical Zod schema in @hashhive/shared. The runtime spec gains a
-// route entry but the schema body stays open — the U4 diff will surface
-// any gaps once the dashboard YAML is compared field-by-field.
-const campaignRowSchema = z.object({}).passthrough().openapi('Campaign')
-const attackRowSchema = z.object({}).passthrough().openapi('Attack')
+// `campaignIdParamSchema`, `campaignRowSchema`, `attackRowSchema`
+// live in `./campaigns-shared.ts` so the lifecycle and attack route
+// modules can reuse them without re-creating an import cycle.
 
 const listCampaignsResponseSchema = z
   .object({
@@ -128,23 +112,7 @@ const updateCampaignResponseSchema = z.object({
   campaign: campaignRowSchema,
 })
 
-const lifecycleResponseSchema = z.object({
-  campaign: campaignRowSchema.nullable(),
-})
-
 const validateCampaignResponseSchema = z.object({}).passthrough()
-
-const listAttacksResponseSchema = z.object({
-  attacks: z.array(attackRowSchema),
-})
-
-const attackResponseSchema = z.object({
-  attack: attackRowSchema,
-})
-
-const deleteAttackResponseSchema = z.object({
-  deleted: z.literal(true),
-})
 
 const listCampaignsRoute = createRoute({
   method: 'get',
@@ -564,247 +532,6 @@ campaignRoutes.openapi(putCampaignRoute, async (c) => {
   return updateCampaignHandler(c, id, data)
 })
 
-// ─── Campaign Lifecycle ─────────────────────────────────────────────
-
-// Action enum matches the spec-named alias routes — `resume` is
-// included alongside `start` even though both map to `running`,
-// because the alias path exposes `/resume` and the parity is
-// documented in the OpenAPI spec.
-const lifecycleSchema = z.object({
-  action: z.enum(['start', 'pause', 'resume', 'stop', 'cancel']),
-})
-
-const lifecycleActionRoute = createRoute({
-  method: 'post',
-  path: '/{id}/lifecycle',
-  tags: ['campaigns'],
-  summary: 'Transition a campaign by action enum (start|pause|resume|stop|cancel)',
-  security: [{ SessionCookie: [] }],
-  middleware: [requireMembershipRole('admin', 'contributor')] as const,
-  request: {
-    params: campaignIdParamSchema,
-    body: {
-      content: {
-        'application/json': { schema: lifecycleSchema },
-      },
-    },
-  },
-  responses: {
-    200: {
-      description: 'Campaign transitioned.',
-      content: { 'application/json': { schema: lifecycleResponseSchema } },
-    },
-    401: sharedResponse(DASHBOARD_RESPONSE_REFS.AuthRequired),
-    403: sharedResponse(DASHBOARD_RESPONSE_REFS.Forbidden),
-    400: sharedResponse(DASHBOARD_RESPONSE_REFS.ValidationFailed),
-    404: sharedResponse(DASHBOARD_RESPONSE_REFS.ResourceNotFound),
-    409: {
-      description: 'Stale state or referenced resources missing.',
-      content: { 'application/json': { schema: z.object({}).passthrough() } },
-    },
-    500: {
-      description: 'Task generation failed during transition.',
-      content: { 'application/json': { schema: z.object({}).passthrough() } },
-    },
-    503: {
-      description: 'Queue or resource lookup unavailable.',
-      content: { 'application/json': { schema: z.object({}).passthrough() } },
-    },
-  },
-})
-
-campaignRoutes.openapi(lifecycleActionRoute, async (c) => {
-  const { id } = c.req.valid('param')
-
-  // transitionCampaign fetches by id alone; without this guard a
-  // contributor in project A could transition a campaign in project
-  // B by guessing the id. Mirrors the per-alias handler check.
-  const { projectId } = c.get('scopedUser')!
-  const existing = await getCampaignById(id)
-  if (!existing || existing.projectId !== projectId) {
-    return dashboardError(c, 404, 'RESOURCE_NOT_FOUND', 'Campaign not found')
-  }
-
-  const { action } = c.req.valid('json')
-
-  const statusMap = {
-    start: 'running',
-    pause: 'paused',
-    resume: 'running',
-    stop: 'draft',
-    cancel: 'cancelled',
-  } as const
-
-  const targetStatus = statusMap[action]
-  const result = await transitionCampaign(id, targetStatus)
-  return respondToTransition(c, result)
-})
-
-// Shared error mapping for transitionCampaign results. Keeps the
-// /lifecycle action-enum route and the spec-named alias routes in
-// lockstep — every recognized service-layer error code maps to the
-// same HTTP status and envelope across both surfaces.
-type TransitionResult = Awaited<ReturnType<typeof transitionCampaign>>
-function respondToTransition(c: Context<AppEnv>, result: TransitionResult) {
-  if ('error' in result) {
-    const code = 'code' in result ? result.code : undefined
-    if (code === 'QUEUE_UNAVAILABLE') {
-      return dashboardError(c, 503, 'SERVICE_UNAVAILABLE', result.error)
-    }
-    if (code === 'RESOURCE_VALIDATION_FAILED') {
-      // Resource lookup itself failed (DB blip). Surface as 503 so
-      // clients can retry rather than treating it as a permanent error.
-      return dashboardError(c, 503, 'SERVICE_UNAVAILABLE', result.error)
-    }
-    if (code === 'RESOURCE_MISSING') {
-      // Precondition failed (referenced resource doesn't exist or
-      // crosses project boundary). 409 Conflict captures this better
-      // than 400 — the request was well-formed, the state isn't.
-      return dashboardError(c, 409, 'RESOURCE_MISSING', result.error)
-    }
-    if (code === 'STALE_STATE') {
-      // Optimistic-concurrency loss: another writer transitioned this
-      // campaign between our read and write. 409 signals the client
-      // should re-fetch and retry against the current state.
-      return dashboardError(c, 409, 'STALE_STATE', result.error)
-    }
-    if (code === 'TASK_GENERATION_FAILED') {
-      return dashboardError(c, 500, 'TASK_GENERATION_FAILED', result.error)
-    }
-    return dashboardError(c, 400, 'INVALID_TRANSITION', result.error)
-  }
-  return c.json({ campaign: result.campaign })
-}
-
-/**
- * Route-layer wrapper around `validateCampaignResources` that turns
- * service throws into the dashboard's `{ error: { code, message } }`
- * envelope. Returns:
- *   - `null` when validation succeeded (caller proceeds)
- *   - a 409 `RESOURCE_MISSING` Response when refs are missing
- *   - a 503 `SERVICE_UNAVAILABLE` Response when the lookup itself
- *     threw (DB blip, query error) — mirrors `transitionCampaign`'s
- *     `RESOURCE_VALIDATION_FAILED` mapping so attack-write paths
- *     have the same retryability semantics as the lifecycle path.
- */
-async function checkResourcesOrErrorResponse(
-  c: Context<AppEnv>,
-  campaign: { projectId: number; hashListId: number | null },
-  attacks: Parameters<typeof validateCampaignResources>[1],
-  logContext: { route: string; campaignId: number; projectId: number }
-): Promise<Response | null> {
-  try {
-    const result = await validateCampaignResources(campaign, attacks)
-    if (result.valid) return null
-    return c.json(
-      {
-        error: {
-          code: 'RESOURCE_MISSING',
-          message: `Referenced resources missing: ${result.missing.join(', ')}`,
-        },
-      },
-      409
-    )
-  } catch (err) {
-    logger.error(
-      { err, ...logContext },
-      'validateCampaignResources threw — surfacing as service unavailable'
-    )
-    return c.json(
-      {
-        error: {
-          code: 'SERVICE_UNAVAILABLE',
-          message: 'Unable to validate campaign resources right now',
-        },
-      },
-      503
-    )
-  }
-}
-
-// Spec-named lifecycle aliases. These delegate to the same
-// transitionCampaign service the action-enum /lifecycle route uses, so
-// behavior (queue check, task generation, event emission, valid-
-// transition allow-list) stays in lockstep. The pre-existing
-// /lifecycle route is kept for the frontend which still calls it.
-const lifecycleAliasStatus = {
-  start: 'running',
-  pause: 'paused',
-  resume: 'running',
-  stop: 'draft',
-  cancel: 'cancelled',
-} as const
-
-type LifecycleAliasAction = keyof typeof lifecycleAliasStatus
-
-async function lifecycleAliasHandler(c: Context<AppEnv>, id: number, action: LifecycleAliasAction) {
-  const { projectId } = c.get('scopedUser')!
-  const existing = await getCampaignById(id)
-  if (!existing || existing.projectId !== projectId) {
-    return dashboardError(c, 404, 'RESOURCE_NOT_FOUND', 'Campaign not found')
-  }
-
-  const result = await transitionCampaign(id, lifecycleAliasStatus[action])
-  return respondToTransition(c, result)
-}
-
-function buildLifecycleAliasRoute(action: LifecycleAliasAction) {
-  return createRoute({
-    method: 'post',
-    path: `/{id}/${action}`,
-    tags: ['campaigns'],
-    summary: `Transition a campaign via the /${action} alias`,
-    security: [{ SessionCookie: [] }],
-    middleware: [requireMembershipRole('admin', 'contributor')] as const,
-    request: {
-      params: campaignIdParamSchema,
-    },
-    responses: {
-      200: {
-        description: 'Campaign transitioned.',
-        content: { 'application/json': { schema: lifecycleResponseSchema } },
-      },
-      401: sharedResponse(DASHBOARD_RESPONSE_REFS.AuthRequired),
-      403: sharedResponse(DASHBOARD_RESPONSE_REFS.Forbidden),
-      400: sharedResponse(DASHBOARD_RESPONSE_REFS.ValidationFailed),
-      404: sharedResponse(DASHBOARD_RESPONSE_REFS.ResourceNotFound),
-      409: {
-        description: 'Stale state or referenced resources missing.',
-        content: { 'application/json': { schema: z.object({}).passthrough() } },
-      },
-      500: {
-        description: 'Task generation failed during transition.',
-        content: { 'application/json': { schema: z.object({}).passthrough() } },
-      },
-      503: {
-        description: 'Queue or resource lookup unavailable.',
-        content: { 'application/json': { schema: z.object({}).passthrough() } },
-      },
-    },
-  })
-}
-
-campaignRoutes.openapi(buildLifecycleAliasRoute('start'), async (c) => {
-  const { id } = c.req.valid('param')
-  return lifecycleAliasHandler(c, id, 'start')
-})
-campaignRoutes.openapi(buildLifecycleAliasRoute('pause'), async (c) => {
-  const { id } = c.req.valid('param')
-  return lifecycleAliasHandler(c, id, 'pause')
-})
-campaignRoutes.openapi(buildLifecycleAliasRoute('resume'), async (c) => {
-  const { id } = c.req.valid('param')
-  return lifecycleAliasHandler(c, id, 'resume')
-})
-campaignRoutes.openapi(buildLifecycleAliasRoute('stop'), async (c) => {
-  const { id } = c.req.valid('param')
-  return lifecycleAliasHandler(c, id, 'stop')
-})
-campaignRoutes.openapi(buildLifecycleAliasRoute('cancel'), async (c) => {
-  const { id } = c.req.valid('param')
-  return lifecycleAliasHandler(c, id, 'cancel')
-})
-
 // ─── DAG Validation ─────────────────────────────────────────────────
 
 const validateCampaignRoute = createRoute({
@@ -848,347 +575,14 @@ campaignRoutes.openapi(validateCampaignRoute, async (c) => {
   return c.json(result, 200)
 })
 
-// ─── Attack Management ──────────────────────────────────────────────
+// ─── Sub-router registrations ──────────────────────────────────────
+//
+// Lifecycle and attack routes live in their own files to keep this
+// module under the 800-line cap. They register against the same
+// `campaignRoutes` router so URL paths and middleware composition
+// stay identical to the single-file form.
 
-const createAttackSchema = z.object({
-  mode: z.number().int().nonnegative(),
-  hashTypeId: z.number().int().positive().optional(),
-  wordlistId: z.number().int().positive().optional(),
-  rulelistId: z.number().int().positive().optional(),
-  masklistId: z.number().int().positive().optional(),
-  advancedConfiguration: z.record(z.string(), z.unknown()).optional(),
-  dependencies: z.array(z.number().int().positive()).optional(),
-})
-
-// Synthetic id used for pre-insert DAG validation. Attack IDs are
-// positive serials, so any negative value is guaranteed not to collide
-// with existing rows.
-const SYNTHETIC_NEW_ATTACK_ID = -1
-
-const createAttackRoute = createRoute({
-  method: 'post',
-  path: '/{id}/attacks',
-  tags: ['campaigns', 'attacks'],
-  summary: 'Create an attack on a draft campaign',
-  security: [{ SessionCookie: [] }],
-  middleware: [requireMembershipRole('admin', 'contributor')] as const,
-  request: {
-    params: campaignIdParamSchema,
-    body: {
-      content: {
-        'application/json': { schema: createAttackSchema },
-      },
-    },
-  },
-  responses: {
-    201: {
-      description: 'Attack created.',
-      content: { 'application/json': { schema: attackResponseSchema } },
-    },
-    401: sharedResponse(DASHBOARD_RESPONSE_REFS.AuthRequired),
-    403: sharedResponse(DASHBOARD_RESPONSE_REFS.Forbidden),
-    400: sharedResponse(DASHBOARD_RESPONSE_REFS.ValidationFailed),
-    404: sharedResponse(DASHBOARD_RESPONSE_REFS.ResourceNotFound),
-    409: {
-      description: 'Referenced resources missing or cross-project.',
-      content: { 'application/json': { schema: z.object({}).passthrough() } },
-    },
-    503: {
-      description: 'Resource validation lookup unavailable.',
-      content: { 'application/json': { schema: z.object({}).passthrough() } },
-    },
-  },
-})
-
-campaignRoutes.openapi(createAttackRoute, async (c) => {
-  const { id: campaignId } = c.req.valid('param')
-
-  // Project-scope guard: requireMembershipRole only validates that the caller
-  // is a contributor *somewhere*; without this check a contributor
-  // in project A could create attacks against a campaign in project
-  // B by guessing the campaign id. 404 on cross-project to avoid
-  // leaking existence.
-  const { projectId } = c.get('scopedUser')!
-  const campaign = await getCampaignById(campaignId)
-  if (!campaign || campaign.projectId !== projectId) {
-    return dashboardError(c, 404, 'RESOURCE_NOT_FOUND', 'Campaign not found')
-  }
-
-  const data = c.req.valid('json')
-
-  // Cross-project resource pre-check at draft write time. FK
-  // constraints only enforce existence; this gate enforces project
-  // ownership so a contributor can't persist attack rows that
-  // reference resources from another project. Skipped when no
-  // nullable resource refs are supplied.
-  const hasResourceRefs =
-    data.hashTypeId != null ||
-    data.wordlistId != null ||
-    data.rulelistId != null ||
-    data.masklistId != null
-  if (hasResourceRefs) {
-    const errResp = await checkResourcesOrErrorResponse(
-      c,
-      { projectId: campaign.projectId, hashListId: null },
-      [
-        {
-          hashTypeId: data.hashTypeId,
-          wordlistId: data.wordlistId,
-          rulelistId: data.rulelistId,
-          masklistId: data.masklistId,
-        },
-      ],
-      { route: 'POST /:id/attacks', campaignId, projectId: campaign.projectId }
-    )
-    if (errResp) return errResp
-  }
-
-  // Pre-insert DAG validation: build the proposed graph (current
-  // attacks + this new attack with a synthetic id) and reject the
-  // request if it would introduce a cycle or reference a missing
-  // attack id. Skipped when no dependencies are supplied — a
-  // dependency-less attack cannot introduce a cycle, and skipping
-  // the listAttacks read keeps the hot path cheap. Mirrors the same
-  // optimization on the PATCH /:id/attacks/:attackId route.
-  if (data.dependencies && data.dependencies.length > 0) {
-    const currentAttacks = await listAttacks(campaignId)
-    const proposed = [
-      ...currentAttacks.map((a) => ({
-        id: a.id,
-        dependencies: a.dependencies as number[] | null,
-      })),
-      {
-        id: SYNTHETIC_NEW_ATTACK_ID,
-        dependencies: data.dependencies,
-      },
-    ]
-    const dagResult = validateProposedDAG(proposed)
-    if (!dagResult.valid) {
-      return dashboardError(c, 400, 'DAG_INVALID', dagResult.error ?? 'Invalid DAG')
-    }
-  }
-
-  const attack = await createAttack({
-    ...data,
-    campaignId,
-    projectId: campaign.projectId,
-  })
-
-  return c.json({ attack }, 201)
-})
-
-const listAttacksRoute = createRoute({
-  method: 'get',
-  path: '/{id}/attacks',
-  tags: ['campaigns', 'attacks'],
-  summary: 'List attacks for a campaign',
-  security: [{ SessionCookie: [] }],
-  middleware: [requireProjectAccess()] as const,
-  request: {
-    params: campaignIdParamSchema,
-  },
-  responses: {
-    200: {
-      description: 'Attacks belonging to the campaign.',
-      content: { 'application/json': { schema: listAttacksResponseSchema } },
-    },
-    401: sharedResponse(DASHBOARD_RESPONSE_REFS.AuthRequired),
-    403: sharedResponse(DASHBOARD_RESPONSE_REFS.Forbidden),
-    400: sharedResponse(DASHBOARD_RESPONSE_REFS.ValidationFailed),
-    404: sharedResponse(DASHBOARD_RESPONSE_REFS.ResourceNotFound),
-  },
-})
-
-campaignRoutes.openapi(listAttacksRoute, async (c) => {
-  const { id: campaignId } = c.req.valid('param')
-  const campaign = await getCampaignById(campaignId)
-
-  // Cross-project enforcement: same reasoning as the /:id/validate
-  // route above -- without the projectId compare, any project member
-  // could enumerate attacks for a campaign in a different project by
-  // guessing the id. 404 to avoid leaking existence.
-  const { projectId } = c.get('scopedUser')!
-  if (!campaign || campaign.projectId !== projectId) {
-    return dashboardError(c, 404, 'RESOURCE_NOT_FOUND', 'Campaign not found')
-  }
-
-  const campaignAttacks = await listAttacks(campaignId)
-  return c.json({ attacks: campaignAttacks }, 200)
-})
-
-const updateAttackSchema = z.object({
-  mode: z.number().int().nonnegative().optional(),
-  hashTypeId: z.number().int().positive().optional(),
-  wordlistId: z.number().int().positive().optional(),
-  rulelistId: z.number().int().positive().optional(),
-  masklistId: z.number().int().positive().optional(),
-  advancedConfiguration: z.record(z.string(), z.unknown()).optional(),
-  dependencies: z.array(z.number().int().positive()).optional(),
-})
-
-const attackPathParamSchema = z.object({
-  id: z.coerce.number().int().positive(),
-  attackId: z.coerce.number().int().positive(),
-})
-
-const updateAttackRoute = createRoute({
-  method: 'patch',
-  path: '/{id}/attacks/{attackId}',
-  tags: ['campaigns', 'attacks'],
-  summary: 'Update an attack on a draft campaign',
-  security: [{ SessionCookie: [] }],
-  middleware: [requireMembershipRole('admin', 'contributor')] as const,
-  request: {
-    params: attackPathParamSchema,
-    body: {
-      content: {
-        'application/json': { schema: updateAttackSchema },
-      },
-    },
-  },
-  responses: {
-    200: {
-      description: 'Attack updated.',
-      content: { 'application/json': { schema: attackResponseSchema } },
-    },
-    401: sharedResponse(DASHBOARD_RESPONSE_REFS.AuthRequired),
-    403: sharedResponse(DASHBOARD_RESPONSE_REFS.Forbidden),
-    400: sharedResponse(DASHBOARD_RESPONSE_REFS.ValidationFailed),
-    404: sharedResponse(DASHBOARD_RESPONSE_REFS.ResourceNotFound),
-    409: {
-      description: 'Referenced resources missing or cross-project.',
-      content: { 'application/json': { schema: z.object({}).passthrough() } },
-    },
-    503: {
-      description: 'Resource validation lookup unavailable.',
-      content: { 'application/json': { schema: z.object({}).passthrough() } },
-    },
-  },
-})
-
-campaignRoutes.openapi(updateAttackRoute, async (c) => {
-  const { id: campaignId, attackId } = c.req.valid('param')
-
-  // Project-scope guard: load the parent campaign unconditionally
-  // (not just when resource refs change) and reject when it doesn't
-  // belong to the caller's active project. Without this gate a
-  // contributor in project A could mutate attacks in project B by
-  // guessing {campaignId, attackId} pairs. The attack-belongs-to-
-  // campaign check below is necessary but not sufficient.
-  const { projectId } = c.get('scopedUser')!
-  const parentCampaign = await getCampaignById(campaignId)
-  if (!parentCampaign || parentCampaign.projectId !== projectId) {
-    return dashboardError(c, 404, 'RESOURCE_NOT_FOUND', 'Campaign not found')
-  }
-
-  // Verify attack belongs to the specified campaign
-  const existing = await getAttackById(attackId)
-  if (!existing || existing.campaignId !== campaignId) {
-    return dashboardError(c, 404, 'RESOURCE_NOT_FOUND', 'Attack not found')
-  }
-
-  const data = c.req.valid('json')
-
-  // Cross-project resource pre-check at draft write time. Only fires
-  // when a resource ref is actually being changed; the existing row
-  // already passed this gate when it was created. Uses the
-  // parentCampaign already loaded above for its projectId.
-  const hasResourceRefChange =
-    data.hashTypeId !== undefined ||
-    data.wordlistId !== undefined ||
-    data.rulelistId !== undefined ||
-    data.masklistId !== undefined
-  if (hasResourceRefChange) {
-    const errResp = await checkResourcesOrErrorResponse(
-      c,
-      { projectId: parentCampaign.projectId, hashListId: null },
-      [
-        {
-          hashTypeId: data.hashTypeId,
-          wordlistId: data.wordlistId,
-          rulelistId: data.rulelistId,
-          masklistId: data.masklistId,
-        },
-      ],
-      { route: 'PATCH /:id/attacks/:attackId', campaignId, projectId: parentCampaign.projectId }
-    )
-    if (errResp) return errResp
-  }
-
-  // Pre-update DAG validation: only when dependencies are being
-  // changed. Other field changes (mode, wordlist, etc.) do not affect
-  // the dependency graph, so skipping the load avoids the extra query.
-  if (data.dependencies !== undefined) {
-    const currentAttacks = await listAttacks(campaignId)
-    const proposed = currentAttacks.map((a) => ({
-      id: a.id,
-      dependencies:
-        a.id === attackId ? (data.dependencies ?? null) : (a.dependencies as number[] | null),
-    }))
-    const dagResult = validateProposedDAG(proposed)
-    if (!dagResult.valid) {
-      return dashboardError(c, 400, 'DAG_INVALID', dagResult.error ?? 'Invalid DAG')
-    }
-  }
-
-  const attack = await updateAttack(attackId, data)
-
-  if (!attack) {
-    return dashboardError(c, 404, 'RESOURCE_NOT_FOUND', 'Attack not found')
-  }
-
-  return c.json({ attack }, 200)
-})
-
-const deleteAttackRoute = createRoute({
-  method: 'delete',
-  path: '/{id}/attacks/{attackId}',
-  tags: ['campaigns', 'attacks'],
-  summary: 'Delete an attack from a draft campaign',
-  security: [{ SessionCookie: [] }],
-  middleware: [requireMembershipRole('admin', 'contributor')] as const,
-  request: {
-    params: attackPathParamSchema,
-  },
-  responses: {
-    200: {
-      description: 'Attack deleted.',
-      content: { 'application/json': { schema: deleteAttackResponseSchema } },
-    },
-    401: sharedResponse(DASHBOARD_RESPONSE_REFS.AuthRequired),
-    403: sharedResponse(DASHBOARD_RESPONSE_REFS.Forbidden),
-    400: sharedResponse(DASHBOARD_RESPONSE_REFS.ValidationFailed),
-    404: sharedResponse(DASHBOARD_RESPONSE_REFS.ResourceNotFound),
-  },
-})
-
-campaignRoutes.openapi(deleteAttackRoute, async (c) => {
-  const { id: campaignId, attackId } = c.req.valid('param')
-  const { projectId } = c.get('scopedUser')!
-
-  // Verify the campaign exists AND belongs to the caller's current
-  // project scope. requireMembershipRole only proves the caller is a
-  // member of the active project; without this check, a user in
-  // project A who knew a valid {campaignId, attackId} pair from
-  // project B could delete project B's attack.
-  const parent = await getCampaignById(campaignId)
-  if (!parent || (projectId !== null && parent.projectId !== projectId)) {
-    return dashboardError(c, 404, 'RESOURCE_NOT_FOUND', 'Campaign not found')
-  }
-
-  // Verify attack belongs to the specified campaign
-  const existing = await getAttackById(attackId)
-  if (!existing || existing.campaignId !== campaignId) {
-    return dashboardError(c, 404, 'RESOURCE_NOT_FOUND', 'Attack not found')
-  }
-
-  const attack = await deleteAttack(attackId)
-
-  if (!attack) {
-    return dashboardError(c, 404, 'RESOURCE_NOT_FOUND', 'Attack not found')
-  }
-
-  return c.json({ deleted: true as const }, 200)
-})
+registerCampaignLifecycleRoutes(campaignRoutes)
+registerCampaignAttackRoutes(campaignRoutes)
 
 export { campaignRoutes }
