+++ packages/backend/src/routes/dashboard/resources.ts
@@ -20,130 +20,33 @@ import {
 } from '../../openapi/components.js'
 import { guessHashType } from '../../services/hash-analysis.js'
 import {
-  abortChunkedUpload,
-  completeChunkedUpload,
   createHashList,
-  createResource,
   deleteHashList,
-  deleteResource,
-  getChunkedUploadStatus,
   getHashItems,
   getHashListById,
   getHashListStats,
-  getResourceById,
   getResourcePresignedUrl,
   importHashList,
-  initiateChunkedUpload,
   listHashLists,
   listHashTypes,
-  listResources,
-  MAX_DIRECT_UPLOAD_BYTES,
-  type ResourceTable,
   ResourceInUseError,
-  uploadChunkPart,
   uploadHashListFile,
   UploadTooLargeError,
-  uploadResourceFile,
 } from '../../services/resources.js'
-
-// Cap the multipart wire body slightly above the direct-upload limit. The
-// extra 1 MB covers multipart overhead (boundaries, field headers) so a
-// genuine 10 MB file isn't rejected by the wire-size check before it
-// reaches the byte-size check in `uploadHashListFile`. Anything larger
-// is rejected before parseBody so the server doesn't buffer GBs into
-// memory. See the multipart branch in POST /hash-lists.
-const MULTIPART_BODY_LIMIT_BYTES = MAX_DIRECT_UPLOAD_BYTES + 1_048_576
-
-/**
- * Reject oversize multipart payloads BEFORE `c.req.parseBody()` buffers
- * the whole body. Two protections:
- *   1. `Transfer-Encoding: chunked` lacks Content-Length, so the cap
- *      below can't enforce. Reject chunked outright (411) and steer
- *      the caller to the streaming chunked-upload endpoint.
- *   2. Declared Content-Length above `MULTIPART_BODY_LIMIT_BYTES` →
- *      413 PAYLOAD_TOO_LARGE.
- * `uploadHashListFile` / `uploadResourceFile` enforce the post-parse
- * byte cap via `UploadTooLargeError` as a backstop. Without this
- * pre-parse guard an authenticated admin/contributor could OOM the
- * backend with a multi-GB body.
- *
- * Returns a Response when the request must be rejected; returns null
- * when the caller should proceed to parseBody().
- */
-function enforceMultipartSizeLimit(c: {
-  req: { header: (k: string) => string | undefined }
-  json: (body: unknown, status: number) => Response
-}): Response | null {
-  const transferEncoding = (c.req.header('transfer-encoding') ?? '').toLowerCase()
-  if (transferEncoding.includes('chunked')) {
-    return c.json(
-      {
-        error: {
-          code: 'LENGTH_REQUIRED',
-          message:
-            'Multipart uploads must include Content-Length. Use the chunked upload endpoint (POST /api/v1/dashboard/resources/upload/initiate) for streamed/large files.',
-        },
-      },
-      411
-    )
-  }
-  const contentLengthRaw = c.req.header('content-length')
-  const contentLength = contentLengthRaw ? Number(contentLengthRaw) : undefined
-  if (
-    typeof contentLength === 'number' &&
-    Number.isFinite(contentLength) &&
-    contentLength > MULTIPART_BODY_LIMIT_BYTES
-  ) {
-    return c.json(
-      {
-        error: {
-          code: 'PAYLOAD_TOO_LARGE',
-          message: `Multipart body (${contentLength} bytes) exceeds ${MULTIPART_BODY_LIMIT_BYTES} bytes. Use the chunked upload endpoint (POST /api/v1/dashboard/resources/upload/initiate) for larger files.`,
-        },
-      },
-      413
-    )
-  }
-  return null
-}
-
-/**
- * Shared query-shape validator for the chunked-upload GET-status /
- * DELETE-abort endpoints. Both routes accept `uploadId` (path) plus
- * `resourceId` (positive integer query) + `resourceType` (one of the
- * known resource buckets) -- truthiness checks alone admitted
- * `resourceId=-1` and arbitrary `resourceType` strings, leaking
- * invalid input into the service layer.
- */
-const RESOURCE_TYPES = ['hash-lists', 'wordlists', 'rulelists', 'masklists'] as const
-const uploadStatusQuerySchema = z.object({
-  resourceId: z.coerce.number().int().positive(),
-  resourceType: z.enum(RESOURCE_TYPES),
-})
-
-// ─── OpenAPI helpers ────────────────────────────────────────────────
-
-// Generic passthrough schema for response shapes that don't have a
-// corresponding shared Zod schema yet. The U4 diff script surfaces
-// these as gaps later — for now they keep the spec parseable without
-// inventing precise types that may drift from the service layer.
-const passthroughObject = (name: string) => z.object({}).passthrough().openapi(name)
-
-const idParamSchema = z.object({
-  id: z.coerce.number().int().positive(),
-})
-
-const uploadIdParamSchema = z.object({
-  uploadId: z.string().min(1),
-})
-
-const uploadPartParamSchema = z.object({
-  uploadId: z.string().min(1),
-  partNumber: z.coerce.number().int().positive(),
-})
-
-const tags: string[] = ['Dashboard/Resources']
-const security = [{ SessionCookie: [] }]
+import { registerChunkedUploadRoutes } from './resources-chunked-upload.js'
+import { registerGenericResourceRoutes } from './resources-generic.js'
+// `enforceMultipartSizeLimit`, `passthroughObject`, `idParamSchema`,
+// `tags`, `security` and the chunked-upload query shape live in
+// `./resources-shared.ts` so the generic-resource factory and
+// chunked-upload route modules can reuse them without re-creating an
+// import cycle through this main router module.
+import {
+  enforceMultipartSizeLimit,
+  idParamSchema,
+  passthroughObject,
+  security,
+  tags,
+} from './resources-shared.js'
 
 const resourceRoutes = new OpenAPIHono<AppEnv>(dashboardOpenApiHonoOptions)
 
@@ -760,563 +663,17 @@ resourceRoutes.openapi(detectHashTypeRoute, async (c) => {
   return c.json({ results }, 200)
 })
 
-// ─── Generic resource routes factory ────────────────────────────────
+// ─── Sub-router registrations ──────────────────────────────────────
+//
+// The wordlists/rulelists/masklists generic factory and the S3
+// multipart chunked-upload session routes live in their own files to
+// keep this module under the 800-line cap. They register against the
+// same `resourceRoutes` router so URL paths and middleware composition
+// stay identical to the single-file form.
 
-function createResourceRoutes(prefix: string, table: ResourceTable) {
-  const createSchema = z.object({
-    name: z.string().min(1).max(255),
-  })
-
-  const listResourceRoute = createRoute({
-    method: 'get',
-    path: `/${prefix}`,
-    tags,
-    summary: `List ${prefix} for the active project`,
-    security,
-    middleware: [requireProjectAccess()] as const,
-    responses: {
-      200: {
-        description: `${prefix} collection`,
-        content: { 'application/json': { schema: passthroughObject(`${prefix}ListResponse`) } },
-      },
-      401: sharedResponse(DASHBOARD_RESPONSE_REFS.AuthRequired),
-      403: sharedResponse(DASHBOARD_RESPONSE_REFS.Forbidden),
-    },
-  })
-
-  resourceRoutes.openapi(listResourceRoute, async (c) => {
-    const { projectId } = c.get('scopedUser')!
-
-    const items = await listResources(table, projectId)
-    return c.json({ [prefix]: items }, 200)
-  })
-
-  const createResourceRoute = createRoute({
-    method: 'post',
-    path: `/${prefix}`,
-    tags,
-    summary: `Create a ${prefix} entry`,
-    security,
-    middleware: [requireMembershipRole('admin', 'contributor')] as const,
-    request: {
-      body: {
-        content: {
-          'application/json': { schema: createSchema },
-        },
-      },
-    },
-    responses: {
-      201: {
-        description: 'Created',
-        content: { 'application/json': { schema: z.object({ item: z.unknown() }) } },
-      },
-      400: sharedResponse(DASHBOARD_RESPONSE_REFS.ValidationFailed),
-      401: sharedResponse(DASHBOARD_RESPONSE_REFS.AuthRequired),
-      403: sharedResponse(DASHBOARD_RESPONSE_REFS.Forbidden),
-    },
-  })
-
-  resourceRoutes.openapi(createResourceRoute, async (c) => {
-    const data = c.req.valid('json')
-    const { projectId } = c.get('scopedUser')!
-    const item = await createResource(table, { ...data, projectId })
-    return c.json({ item }, 201)
-  })
-
-  const getResourceRoute = createRoute({
-    method: 'get',
-    path: `/${prefix}/{id}`,
-    tags,
-    summary: `Get a ${prefix} entry by id`,
-    security,
-    middleware: [requireProjectAccess()] as const,
-    request: { params: idParamSchema },
-    responses: {
-      200: {
-        description: 'Resource',
-        content: { 'application/json': { schema: z.object({ item: z.unknown() }) } },
-      },
-      401: sharedResponse(DASHBOARD_RESPONSE_REFS.AuthRequired),
-      403: sharedResponse(DASHBOARD_RESPONSE_REFS.Forbidden),
-      404: sharedResponse(DASHBOARD_RESPONSE_REFS.ResourceNotFound),
-    },
-  })
-
-  resourceRoutes.openapi(getResourceRoute, async (c) => {
-    const { projectId } = c.get('scopedUser')!
-    const id = Number(c.req.param('id'))
-    const item = await getResourceById(table, id, projectId)
-
-    if (!item) {
-      return dashboardError(c, 404, 'RESOURCE_NOT_FOUND', `${prefix} item not found`)
-    }
-
-    return c.json({ item }, 200)
-  })
-
-  const uploadResourceRoute = createRoute({
-    method: 'post',
-    path: `/${prefix}/{id}/upload`,
-    tags,
-    summary: `Upload the file payload for a ${prefix} entry`,
-    security,
-    middleware: [requireMembershipRole('admin', 'contributor')] as const,
-    request: {
-      params: idParamSchema,
-      body: {
-        content: {
-          'multipart/form-data': {
-            schema: z.object({ file: z.unknown() }),
-          },
-        },
-      },
-    },
-    responses: {
-      200: {
-        description: 'Upload accepted',
-        content: {
-          'application/json': { schema: passthroughObject(`${prefix}UploadResult`) },
-        },
-      },
-      400: sharedResponse(DASHBOARD_RESPONSE_REFS.ValidationFailed),
-      401: sharedResponse(DASHBOARD_RESPONSE_REFS.AuthRequired),
-      403: sharedResponse(DASHBOARD_RESPONSE_REFS.Forbidden),
-      404: sharedResponse(DASHBOARD_RESPONSE_REFS.ResourceNotFound),
-      411: {
-        description: 'Length Required (chunked transfer-encoding rejected)',
-        content: { 'application/json': { schema: passthroughObject('LengthRequiredError') } },
-      },
-      413: {
-        description: 'Payload Too Large',
-        content: { 'application/json': { schema: passthroughObject('PayloadTooLargeError') } },
-      },
-    },
-  })
-
-  resourceRoutes.openapi(uploadResourceRoute, async (c) => {
-    const { projectId } = c.get('scopedUser')!
-    const id = Number(c.req.param('id'))
-    const item = await getResourceById(table, id, projectId)
-
-    if (!item) {
-      return dashboardError(c, 404, 'RESOURCE_NOT_FOUND', `${prefix} item not found`)
-    }
-
-    const sizeGuardResponse = enforceMultipartSizeLimit(c)
-    if (sizeGuardResponse) return sizeGuardResponse
-
-    const body = await c.req.parseBody()
-    const file = body['file']
-
-    if (!(file instanceof File)) {
-      return dashboardError(c, 400, 'VALIDATION_ERROR', 'file field is required')
-    }
-
-    try {
-      const result = await uploadResourceFile(table, id, projectId, prefix, file)
-      return c.json(result, 200)
-    } catch (err) {
-      if (err instanceof UploadTooLargeError) {
-        return c.json(
-          {
-            error: {
-              code: 'PAYLOAD_TOO_LARGE',
-              message: `File size (${err.size} bytes) exceeds the direct-upload limit (${err.limit} bytes). Use the chunked upload endpoint (POST /api/v1/dashboard/resources/upload/initiate) for larger files.`,
-            },
-          },
-          413
-        )
-      }
-      throw err
-    }
-  })
-
-  const deleteResourceRoute = createRoute({
-    method: 'delete',
-    path: `/${prefix}/{id}`,
-    tags,
-    summary: `Delete a ${prefix} entry`,
-    security,
-    middleware: [requireMembershipRole('admin', 'contributor')] as const,
-    request: { params: idParamSchema },
-    responses: {
-      204: { description: 'Deleted' },
-      400: sharedResponse(DASHBOARD_RESPONSE_REFS.ValidationFailed),
-      401: sharedResponse(DASHBOARD_RESPONSE_REFS.AuthRequired),
-      403: sharedResponse(DASHBOARD_RESPONSE_REFS.Forbidden),
-      404: sharedResponse(DASHBOARD_RESPONSE_REFS.ResourceNotFound),
-      409: {
-        description: 'Resource in use',
-        content: { 'application/json': { schema: passthroughObject('ResourceInUseError') } },
-      },
-    },
-  })
-
-  resourceRoutes.openapi(deleteResourceRoute, async (c) => {
-    const { projectId } = c.get('scopedUser')!
-    const id = Number(c.req.param('id'))
-    if (!Number.isInteger(id) || id <= 0) {
-      return dashboardError(c, 400, 'VALIDATION_ERROR', `Invalid ${prefix} id`)
-    }
-    try {
-      const deleted = await deleteResource(table, id, projectId, prefix)
-      if (!deleted) {
-        return dashboardError(c, 404, 'RESOURCE_NOT_FOUND', `${prefix} item not found`)
-      }
-      return c.body(null, 204)
-    } catch (err) {
-      if (err instanceof ResourceInUseError) {
-        return dashboardError(c, 409, 'RESOURCE_IN_USE', err.message)
-      }
-      throw err
-    }
-  })
-
-  const downloadResourceRoute = createRoute({
-    method: 'get',
-    path: `/${prefix}/{id}/download`,
-    tags,
-    summary: `Get a presigned download URL for a ${prefix} entry`,
-    security,
-    middleware: [requireProjectAccess()] as const,
-    request: { params: idParamSchema },
-    responses: {
-      200: {
-        description: 'Presigned URL',
-        content: { 'application/json': { schema: z.object({ url: z.string() }) } },
-      },
-      400: sharedResponse(DASHBOARD_RESPONSE_REFS.ValidationFailed),
-      401: sharedResponse(DASHBOARD_RESPONSE_REFS.AuthRequired),
-      403: sharedResponse(DASHBOARD_RESPONSE_REFS.Forbidden),
-      404: sharedResponse(DASHBOARD_RESPONSE_REFS.ResourceNotFound),
-    },
-  })
-
-  resourceRoutes.openapi(downloadResourceRoute, async (c) => {
-    const { projectId } = c.get('scopedUser')!
-    const id = Number(c.req.param('id'))
-    const item = await getResourceById(table, id, projectId)
-
-    if (!item) {
-      return dashboardError(c, 404, 'RESOURCE_NOT_FOUND', `${prefix} item not found`)
-    }
-
-    const fileRef = (item as Record<string, unknown>)['fileRef'] as {
-      bucket?: string
-      key?: string
-      name?: string
-    } | null
-    if (!fileRef?.bucket || !fileRef?.key) {
-      return dashboardError(c, 400, 'VALIDATION_ERROR', `${prefix} item has no uploaded file`)
-    }
-
-    const url = await getResourcePresignedUrl({
-      bucket: fileRef.bucket,
-      key: fileRef.key,
-      ...(fileRef.name ? { name: fileRef.name } : {}),
-    })
-    return c.json({ url }, 200)
-  })
-}
-
-createResourceRoutes('wordlists', wordLists)
-createResourceRoutes('rulelists', ruleLists)
-createResourceRoutes('masklists', maskLists)
-
-// ─── Chunked Upload (S3 Multipart) ──────────────────────────────────
-// These endpoints do NOT use zValidator for the body on PUT because
-// the request body is raw binary data (not JSON). Body-parsing
-// middleware would consume the stream before we can forward it to S3.
-
-const initiateUploadSchema = z.object({
-  resourceType: z.enum(['hash-lists', 'wordlists', 'rulelists', 'masklists']),
-  name: z.string().min(1).max(255),
-  fileSize: z.number().int().positive().max(500_000_000_000),
-  contentType: z.string().optional(),
-})
-
-const initiateUploadRoute = createRoute({
-  method: 'post',
-  path: '/upload/initiate',
-  tags,
-  summary: 'Initiate an S3 multipart chunked upload session',
-  security,
-  middleware: [requireMembershipRole('admin', 'contributor')] as const,
-  request: {
-    body: {
-      content: {
-        'application/json': { schema: initiateUploadSchema },
-      },
-    },
-  },
-  responses: {
-    201: {
-      description: 'Upload session initiated',
-      content: { 'application/json': { schema: passthroughObject('UploadInitiateResult') } },
-    },
-    400: sharedResponse(DASHBOARD_RESPONSE_REFS.ValidationFailed),
-    401: sharedResponse(DASHBOARD_RESPONSE_REFS.AuthRequired),
-    403: sharedResponse(DASHBOARD_RESPONSE_REFS.Forbidden),
-    500: {
-      description: 'Upload initiation failed',
-      content: { 'application/json': { schema: passthroughObject('UploadInitFailedError') } },
-    },
-  },
-})
-
-resourceRoutes.openapi(initiateUploadRoute, async (c) => {
-  const data = c.req.valid('json')
-  const { projectId } = c.get('scopedUser')!
-
-  try {
-    const result = await initiateChunkedUpload({ ...data, projectId })
-    return c.json(result, 201)
-  } catch (err) {
-    logger.error({ err }, 'Failed to initiate chunked upload')
-    return dashboardError(c, 500, 'UPLOAD_INIT_FAILED', 'Failed to initiate upload')
-  }
-})
-
-const uploadPartQuerySchema = z.object({
-  resourceId: z.coerce.number().int().positive(),
-  resourceType: z.enum(RESOURCE_TYPES),
-})
-
-const uploadPartRoute = createRoute({
-  method: 'put',
-  path: '/upload/{uploadId}/part/{partNumber}',
-  tags,
-  summary: 'Upload a single chunk part for a multipart upload session',
-  security,
-  middleware: [requireMembershipRole('admin', 'contributor')] as const,
-  request: {
-    params: uploadPartParamSchema,
-    query: uploadPartQuerySchema,
-    body: {
-      content: {
-        'application/octet-stream': { schema: z.unknown() },
-      },
-    },
-  },
-  responses: {
-    200: {
-      description: 'Part uploaded',
-      content: { 'application/json': { schema: passthroughObject('UploadPartResult') } },
-    },
-    400: sharedResponse(DASHBOARD_RESPONSE_REFS.ValidationFailed),
-    401: sharedResponse(DASHBOARD_RESPONSE_REFS.AuthRequired),
-    403: sharedResponse(DASHBOARD_RESPONSE_REFS.Forbidden),
-    500: {
-      description: 'Upload part failed',
-      content: { 'application/json': { schema: passthroughObject('UploadPartFailedError') } },
-    },
-  },
-})
-
-resourceRoutes.openapi(uploadPartRoute, async (c) => {
-  const uploadId = c.req.param('uploadId')
-  const partNumber = Number(c.req.param('partNumber'))
-  const resourceId = Number(c.req.query('resourceId'))
-  const resourceType = c.req.query('resourceType')
-
-  if (!uploadId || !partNumber || !resourceId || !resourceType) {
-    return c.json(
-      {
-        error: {
-          code: 'VALIDATION_ERROR',
-          message: 'uploadId, partNumber, resourceId, and resourceType are required',
-        },
-      },
-      400
-    )
-  }
-
-  // Read the raw body as a Uint8Array — do NOT use c.req.json() or c.req.parseBody()
-  const body = await c.req.arrayBuffer()
-  const chunk = new Uint8Array(body)
-
-  if (chunk.byteLength === 0) {
-    return dashboardError(c, 400, 'VALIDATION_ERROR', 'Request body is empty')
-  }
-
-  const { projectId } = c.get('scopedUser')!
-
-  try {
-    const result = await uploadChunkPart(
-      uploadId,
-      partNumber,
-      chunk,
-      resourceId,
-      resourceType,
-      projectId
-    )
-    return c.json(result, 200)
-  } catch (err) {
-    logger.error({ err, uploadId, partNumber }, 'Failed to upload part')
-    return dashboardError(c, 500, 'UPLOAD_PART_FAILED', 'Failed to upload part')
-  }
-})
-
-const completeUploadSchema = z.object({
-  parts: z
-    .array(
-      z.object({
-        partNumber: z.number().int().positive(),
-        etag: z.string().min(1),
-      })
-    )
-    .min(1),
-  resourceId: z.number().int().positive(),
-  resourceType: z.enum(['hash-lists', 'wordlists', 'rulelists', 'masklists']),
-})
-
-const completeUploadRoute = createRoute({
-  method: 'post',
-  path: '/upload/{uploadId}/complete',
-  tags,
-  summary: 'Complete a multipart upload session',
-  security,
-  middleware: [requireMembershipRole('admin', 'contributor')] as const,
-  request: {
-    params: uploadIdParamSchema,
-    body: {
-      content: {
-        'application/json': { schema: completeUploadSchema },
-      },
-    },
-  },
-  responses: {
-    200: {
-      description: 'Upload completed',
-      content: { 'application/json': { schema: passthroughObject('UploadCompleteResult') } },
-    },
-    400: sharedResponse(DASHBOARD_RESPONSE_REFS.ValidationFailed),
-    401: sharedResponse(DASHBOARD_RESPONSE_REFS.AuthRequired),
-    403: sharedResponse(DASHBOARD_RESPONSE_REFS.Forbidden),
-    500: {
-      description: 'Upload completion failed',
-      content: { 'application/json': { schema: passthroughObject('UploadCompleteFailedError') } },
-    },
-  },
-})
-
-resourceRoutes.openapi(completeUploadRoute, async (c) => {
-  const uploadId = c.req.param('uploadId')
-  const { parts, resourceId, resourceType } = c.req.valid('json')
-  const { projectId } = c.get('scopedUser')!
-
-  try {
-    const result = await completeChunkedUpload(uploadId, parts, resourceId, resourceType, projectId)
-    return c.json(result, 200)
-  } catch (err) {
-    logger.error({ err, uploadId }, 'Failed to complete chunked upload')
-    return dashboardError(c, 500, 'UPLOAD_COMPLETE_FAILED', 'Failed to complete upload')
-  }
-})
-
-const abortUploadRoute = createRoute({
-  method: 'delete',
-  path: '/upload/{uploadId}',
-  tags,
-  summary: 'Abort an in-progress multipart upload session',
-  security,
-  middleware: [requireMembershipRole('admin', 'contributor')] as const,
-  request: {
-    params: uploadIdParamSchema,
-    query: uploadStatusQuerySchema,
-  },
-  responses: {
-    200: {
-      description: 'Abort acknowledged',
-      content: { 'application/json': { schema: z.object({ acknowledged: z.boolean() }) } },
-    },
-    400: sharedResponse(DASHBOARD_RESPONSE_REFS.ValidationFailed),
-    401: sharedResponse(DASHBOARD_RESPONSE_REFS.AuthRequired),
-    403: sharedResponse(DASHBOARD_RESPONSE_REFS.Forbidden),
-  },
-})
-
-resourceRoutes.openapi(abortUploadRoute, async (c) => {
-  const uploadId = c.req.param('uploadId')
-  if (!uploadId) {
-    return dashboardError(c, 400, 'VALIDATION_ERROR', 'uploadId is required')
-  }
-  const parsed = uploadStatusQuerySchema.safeParse({
-    resourceId: c.req.query('resourceId'),
-    resourceType: c.req.query('resourceType'),
-  })
-  if (!parsed.success) {
-    return c.json(
-      {
-        error: {
-          code: 'VALIDATION_ERROR',
-          message: parsed.error.issues.map((i) => i.message).join('; '),
-        },
-      },
-      400
-    )
-  }
-  const { resourceId, resourceType } = parsed.data
-
-  const { projectId } = c.get('scopedUser')!
-
-  await abortChunkedUpload(uploadId, resourceId, resourceType, projectId)
-  return c.json({ acknowledged: true }, 200)
-})
-
-const uploadStatusRoute = createRoute({
-  method: 'get',
-  path: '/upload/{uploadId}/status',
-  tags,
-  summary: 'Get the status of an in-progress multipart upload session',
-  security,
-  middleware: [requireMembershipRole('admin', 'contributor')] as const,
-  request: {
-    params: uploadIdParamSchema,
-    query: uploadStatusQuerySchema,
-  },
-  responses: {
-    200: {
-      description: 'Upload status',
-      content: { 'application/json': { schema: passthroughObject('UploadStatusResult') } },
-    },
-    400: sharedResponse(DASHBOARD_RESPONSE_REFS.ValidationFailed),
-    401: sharedResponse(DASHBOARD_RESPONSE_REFS.AuthRequired),
-    403: sharedResponse(DASHBOARD_RESPONSE_REFS.Forbidden),
-    404: sharedResponse(DASHBOARD_RESPONSE_REFS.ResourceNotFound),
-  },
-})
-
-resourceRoutes.openapi(uploadStatusRoute, async (c) => {
-  const uploadId = c.req.param('uploadId')
-  if (!uploadId) {
-    return dashboardError(c, 400, 'VALIDATION_ERROR', 'uploadId is required')
-  }
-  const parsed = uploadStatusQuerySchema.safeParse({
-    resourceId: c.req.query('resourceId'),
-    resourceType: c.req.query('resourceType'),
-  })
-  if (!parsed.success) {
-    return c.json(
-      {
-        error: {
-          code: 'VALIDATION_ERROR',
-          message: parsed.error.issues.map((i) => i.message).join('; '),
-        },
-      },
-      400
-    )
-  }
-  const { resourceId, resourceType } = parsed.data
-
-  const { projectId } = c.get('scopedUser')!
-
-  const result = await getChunkedUploadStatus(uploadId, resourceId, resourceType, projectId)
-  if (!result) {
-    return dashboardError(c, 404, 'RESOURCE_NOT_FOUND', 'Upload not found')
-  }
-
-  return c.json({ uploadId, ...result }, 200)
-})
+registerGenericResourceRoutes(resourceRoutes, 'wordlists', wordLists)
+registerGenericResourceRoutes(resourceRoutes, 'rulelists', ruleLists)
+registerGenericResourceRoutes(resourceRoutes, 'masklists', maskLists)
+registerChunkedUploadRoutes(resourceRoutes)
 
 export { resourceRoutes }
